+++ tests/Traits/PdfRoundedRectangleTraitTest.php
@@ -22,12 +22,39 @@ use PHPUnit\Framework\TestCase;
 
 final class PdfRoundedRectangleTraitTest extends TestCase
 {
+    public function testMoveBelow(): void
+    {
+        $document = new PdfDocumentRoundedRectangle();
+        $document->addPage();
+        $document->roundedRect(100, 50, 5, 10, 20, move: PdfMove::BELOW);
+        self::assertSame(10.0, $document->getX());
+        self::assertSame(70.0, $document->getY());
+    }
+
+    public function testMoveNewLine(): void
+    {
+        $document = new PdfDocumentRoundedRectangle();
+        $document->addPage();
+        $document->roundedRect(100, 50, 5, 10, 20, move: PdfMove::NEW_LINE);
+        self::assertSame($document->getLeftMargin(), $document->getX());
+        self::assertSame(70.0, $document->getY());
+    }
+
+    public function testMoveRight(): void
+    {
+        $document = new PdfDocumentRoundedRectangle();
+        $document->addPage();
+        $document->roundedRect(100, 50, 5, 10, 20);
+        self::assertSame(110.0, $document->getX());
+        self::assertSame(20.0, $document->getY());
+    }
+
     public function testRoundedInvalidRadius(): void
     {
         $document = new PdfDocumentRoundedRectangle();
         self::expectException(PdfException::class);
         self::expectExceptionMessage('Invalid radius: 40, maximum allowed: 25.');
-        $document->roundedRect(0, 0, 50, 100, 40);
+        $document->roundedRect(50, 100, 40);
     }
 
     public function testRoundedRadiusNotPositive(): void
@@ -35,7 +62,7 @@ final class PdfRoundedRectangleTraitTest extends TestCase
         $document = new PdfDocumentRoundedRectangle();
         self::expectException(PdfException::class);
         self::expectExceptionMessage('The radius must be positive, 0 given.');
-        $document->roundedRect(0, 0, 50, 100, 0);
+        $document->roundedRect(50, 100, 0);
     }
 
     public function testRoundedRect(): void
@@ -45,30 +72,26 @@ final class PdfRoundedRectangleTraitTest extends TestCase
         $document->setFillColor(PdfRgbColor::darkGray());
         $document->setDrawColor(PdfRgbColor::red());
 
-        $x = $document->getX();
-        $y = $document->getY();
+        $document->setXY(20, 20);
         $document->setLineWidth(1.5);
         $document->roundedRect(
-            x: $x,
-            y: $y,
             width: 20,
             height: 20,
             radius: 5,
             move: PdfMove::BELOW
         );
+        self::assertSame(20.0, $document->getX());
+        self::assertSame(40.0, $document->getY());
 
-        $x = $document->getX();
-        $y = $document->getY() + 10;
         $document->setLineWidth(0.5);
         $document->roundedRect(
-            x: $x,
-            y: $y,
             width: 100,
             height: 10,
             radius: 5,
-            move: PdfMove::NEW_LINE
+            move: PdfMove::BELOW
         );
-        self::assertSame(1, $document->getPage());
+        self::assertSame(20.0, $document->getX());
+        self::assertSame(50.0, $document->getY());
     }
 
     public function testRoundedRectangle(): void
@@ -77,6 +100,7 @@ final class PdfRoundedRectangleTraitTest extends TestCase
         $document = new PdfDocumentRoundedRectangle();
         $document->addPage();
         $document->roundedRectangle($rect, 5);
-        self::assertSame(1, $document->getPage());
+        self::assertSame($rect->right(), $document->getX());
+        self::assertSame($rect->y, $document->getY());
     }
 }